Fixes for getConfigurations
For some cameras, the 'name' attribute is an array (Pelco)
Now gets the 'token', panTiltLimits, and zoomLimits.
Multiple checks added for validity of response.
Gets rid of "TypeError: Cannot read property '0' of undefined" fault parsing on Pelco as Pelco does not put the 'zoom' attribute at all in the defaultPTZSpeed object.