Remove subscriptionReference check in PullMessageResponse
The ONVIF Core Spec is a bit vague but from what I can understand an Event message
that comes back from a PullMessage command does
not need to include a subscriptionReference (and HikVision don't send it)

If the library used the other events mechanism, where
the NVT sends a Notity message to this llibrary then we would check the subscriptionReference.
But this library does not support that method of events.

Thanks to
Michiel Mooring @Gielert who reported the issue
Nick Whyte @nickw444 - who has used this in his fork