Fix SOAPAuth timeShift if the local computer's time changes. eg when NTP adjusts the time
The Raspberry Pi does not have a battery backed clock (although upgrade kits are avaiable).
On a standard Pi the OS boots with a fake time based on the time the OS was last shut down.Then some time later the correct time may be received via NTP.
This causes a problem for the 'timeShift' code in this library, needed for correct SOAP Authentication headers.
So switch to using a local clock that does not change. I use process uptime but kernel uptime would work too