Fix -  Wasstarting mock server when testing real device
the enviroment vars HOSTNAME, PORT, USERNAME, PASSWORD are used to point
the mocha tests at a real device.
However bugs had snuck in and when testing with a real device we were still starting the
Mock Server.
So, only start the MockServer when there is no HOSTNAME et.
Also required mocha to be bumped for the mocha.opts file and --exclude
so mocha did not try and test (and run) the mockserver.
I just chose mocha 6, no real reason other than I know it has --exclude in mocha.opts