Add Media2/Profile T getProfiles XML
If the NVT supports Media2 API, use it to get a list of profiles.
This allows us to use H265 video.

The XML returned from Media2 is slightly different to the reply from Media(1).
So adapt the Media2 reply so the callback returns data that is in the
same format as Media1 so we don't need to change the test of this library.
Tested with 2 different HikVision cameras.