Add ContinuousMove onlySendXY and onlySendZ to support Sony SRG-XP1
This camera does not like X, Y and Z all sent in one command.
We must send just XY (pan/tilt) or Z (zoom) but not both.

My hunch is that the camera uses VISCA behind the scenes
and VISCA has a PAN/TILT command and different command for ZOOM.