Handle Hik camera that says 'unauthorized' and not 'sender not authorized'
The root problem was that this library determined if there was an authentication
error by looking for the ASCII String ''sender not authorized' and did not look at
the HTTP Status Code for a 401 result.
The HikVision camera returned some HTML with 'unauthorized' and not 'sender not authorized'

The library now captures the HTTP Status Code (eg 401) and passes it back to
the user of the library with an extra parameter after the (err,xml,data) parameters
of all the ONVIF command callbacks.
It also adds the StatusCode to the RawResult Event

No the library can use the HTTP Status Code, use it in the getSystemDateAndTime code.

This fixes connection to a HikVision camera that required authoriztion for GetSystemDateAndTime
but the library then failed to send authentication due to a bug.

Sorry, VSCode reformatted some of the files so there are whitespace
changes in here too