workaround buggy D-Link DCS-8635LH
Three different bugs found with the camera.
1) I connect to target camera over a tunnel. I connect to 127.0.0.1
The camera on the remote network is 192.168.99.200
ONVIF Replies in GetCapabilities tell me the camera URL
but they claim the camera is http://127.0.0.1.99.200
A mashup of the IPs.
A possible buffer overflow issue here.

2) Using Media2 GetProfiles with WS-Security fails. Camera reports this error
The data in element 'Security' must be understood
(If we have an error we now fall back to Media1, like ODM uses)

3) Probably related to #1 but accessing the camera via ONVIF APIs
can crash the camera's ONVIF stack requiring a reboot, probably due
to a buffer overflow.