Events - fix terminationTime
The Pull Request worker loop was not waiting for terminationTime to be updated (by RENEW) before going around the loop again.
This meant the terminationTime was wrong and the code re-subscribed.

Add code to set the Read Timeout on the Socket.
Add timeout of PT1M to the Pull request (fixes Vista camera)
As the pull reqests now have a 1 Min timeout, make the Subsription
and Renew ask for a 2 minute subscription

Also remove the 30 second timeout.